+++ ICD.Connect.Conferencing.Zoom/Responses/AbstractZoomRoomResponse.cs
@@ -6,6 +6,14 @@ namespace ICD.Connect.Conferencing.Zoom.Responses
 {
 	public abstract class AbstractZoomRoomResponse
 	{
+		private static readonly JsonSerializerSettings s_Settings =
+			new JsonSerializerSettings
+		{
+#if !SIMPLSHARP
+			DateParseHandling = DateParseHandling.None
+#endif
+		};
+
 		[CanBeNull]
 		public static AbstractZoomRoomResponse DeserializeResponse(string data, out AttributeKey key)
 		{
@@ -16,7 +24,7 @@ namespace ICD.Connect.Conferencing.Zoom.Responses
 			Type responseType = key.GetResponseType();
 			return responseType == null
 				? null
-				: JsonConvert.DeserializeObject(data, responseType) as AbstractZoomRoomResponse;
+				: JsonConvert.DeserializeObject(data, responseType, s_Settings) as AbstractZoomRoomResponse;
 		}
 	}
 }